Started duplicate code reduction and added workspaces for shared libraries
