+++ Cargo.toml
@@ -1,0 +1,18 @@
+[workspace]
+
+members = ["contracts/cw721-marketplace", "contracts/cw721-marketplace-single-collection", "contracts/cw721-marketplace-permissioned", "utils"]
+
+[workspace.dependencies]
+cosmwasm-schema = "~1.0.0"
+cw-multi-test = "0.13.4"
+cosmwasm-std = "~1.5.0"
+cosmwasm-storage = "~1.0.0"
+cw-storage-plus = "0.12"
+cw2 = "0.12"
+cw20 = { version = "0.13.4" }
+cw20-base = "0.13.4"
+cw721-base = "0.13.4"
+cw721 = "0.13.4"
+schemars = "0.8"
+serde = { version = "1.0", default-features = false, features = ["derive"] }
+thiserror = "1.0"
