+++ contracts/cw721-marketplace-single-collection/src/contract.rs
@@ -13,8 +13,9 @@ use crate::query::{
     query_swaps_by_denom, query_swaps_by_payment_type, query_swaps_by_price, query_swaps_of_token,
 };
 use crate::msg::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
-use crate::state::{Config, CONFIG, SwapType};
+use crate::state::{Config, CONFIG};
 use crate::error::ContractError;
+use utils::prelude::SwapType;
 
 use cw2::{get_contract_version, set_contract_version};
