+++ contracts/cw721-marketplace-single-collection/src/state.rs
@@ -9,6 +9,7 @@ use cw_storage_plus::{
 };
 
 use cw20::{Expiration};
+use utils::prelude::CW721Swap;
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct Config {
@@ -16,23 +17,6 @@ pub struct Config {
     pub denom: String,
     pub cw721: Addr,
     pub fees: u64,
-}
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub enum SwapType {
-    Offer,
-    Sale
-}
-// swap type of false equals offer, swap type of true equals buy
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct CW721Swap {
-    pub id: String,
-    pub creator: Addr,
-    pub nft_contract: Addr,
-    pub payment_token: Option<Addr>,
-    pub token_id: String,
-    pub expires: Expiration,
-    pub price: Uint128,
-    pub swap_type: SwapType,
 }
 
 pub fn all_swap_ids<'a>(
@@ -46,11 +30,6 @@ pub fn all_swap_ids<'a>(
         .collect()
 }
 
-impl CW721Swap {
-    pub fn is_expired(&self, block: &BlockInfo) -> bool {
-        self.expires.is_expired(block)
-    }
-}
 
 pub const SWAPS: Map<&str, CW721Swap> = Map::new("cw721_swap");
 pub const CONFIG: Item<Config> = Item::new("config");
