+++ utils/src/swap.rs
@@ -1,0 +1,28 @@
+use cosmwasm_std::{Addr, BlockInfo, Uint128};
+use cw20::Expiration;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub enum SwapType {
+    Offer,
+    Sale
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct CW721Swap {
+    pub id: String,
+    pub creator: Addr,
+    pub nft_contract: Addr,
+    pub payment_token: Option<Addr>,
+    pub token_id: String,
+    pub expires: Expiration,
+    pub price: Uint128,
+    pub swap_type: SwapType,
+}
+
+impl CW721Swap {
+    pub fn is_expired(&self, block: &BlockInfo) -> bool {
+        self.expires.is_expired(block)
+    }
+}
